404 error handler

+++ src/main/java/br/dev/rafael/cyberspace/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.dev.rafael.cyberspace.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -58,23 +59,35 @@ public class TopicosController {
 	
 	@GetMapping("/{id}")
 	@Transactional
-	public DetailTopicoDto detail(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetailTopicoDto(topico);
+	public ResponseEntity<DetailTopicoDto> detail(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+			return ResponseEntity.ok(new DetailTopicoDto(topico.get()));		
+		}
+		return ResponseEntity.notFound().build();		
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> update(@PathVariable Long id, @RequestBody @Valid UpdateTopicoForm form){
-		Topico topico = form.update(id, topicoRepository);
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = form.update(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));	
+		}
+		return ResponseEntity.notFound().build();
+		
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id){
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 }
